+++ packages/skeleton-react/src/components/Toast/create-toaster.ts
@@ -1,0 +1,5 @@
+import * as toast from '@zag-js/toast';
+
+export function createToaster(options: toast.StoreProps = {}) {
+	return toast.createStore(options);
+}